delete method refactoring

+++ src/app/components/books-card/books-card.component.ts
@@ -12,19 +12,19 @@ export class BooksCardComponent {
   get books$() {
     return this.booksService.booksSource;
   }
-
   constructor(private booksService: BooksService) {}
 
   addFavorite(book: BookDetail): void {
-    book.isFavorite = true;
-    this.booksService.saveFavorite(book);
-  };
 
-  removeFavorite(book: BookDetail) {
-    book.isFavorite = false;
+    if(!book.isFavorite){
+      book.isFavorite = true;
+      this.booksService.saveFavorite(book);
+    } else {
+      book.isFavorite = false;
       this.booksService.removeFavorite(book);
       this.booksService.addFavoriteBooks();
-  }
+    }
+  };
 
   ngOnInit(): void {
   }
